refactor(*): replace relative import to absolute using webpack alias
