+++ web-components/src/components/alert-banner/AlertBanner.stories.ts
@@ -1,7 +1,7 @@
 import { withA11y } from "@storybook/addon-a11y";
 import { boolean, select, text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-html";
-import "../icon/Icon";
+import "@/components/icon/Icon";
 import "./AlertBanner";
 
 export default {
