+++ web-components/src/components/avatar/Avatar.ts
@@ -6,8 +6,8 @@
  *
  */
 
-import "../icon/Icon";
-import "../loading/Loading";
+import "@/components/icon/Icon";
+import "@/components/loading/Loading";
 import reset from "@/wc_scss/reset.scss";
 import { customElement, html, internalProperty, LitElement, property } from "lit-element";
 import { nothing } from "lit-html";
