+++ web-components/src/components/button/Button.stories.ts
@@ -1,10 +1,10 @@
-import "./Button";
-import "../icon/Icon";
-import "../theme/Theme";
+import "@/components/button/Button";
+import "@/components/icon/Icon";
+import "@/components/theme/Theme";
 import { withA11y } from "@storybook/addon-a11y";
 import { boolean, select, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-element";
-import { buttonColor, buttonSize, buttonTag, buttonType, buttonVariant } from "./Button";
+import { buttonColor, buttonSize, buttonTag, buttonType, buttonVariant } from "@/components/button/Button";
 
 export default {
   title: "Button",
