+++ web-components/src/components/button/Button.ts
@@ -6,9 +6,9 @@
  *
  */
 
-import activityButtonStyles from "../activity-button/scss/module.scss";
-import "../loading/Loading";
-import "../spinner/Spinner";
+import activityButtonStyles from "@/components/activity-button/scss/module.scss";
+import "@/components/loading/Loading";
+import "@/components/spinner/Spinner";
 import { Key } from "@/constants";
 import reset from "@/wc_scss/reset.scss";
 import { customElement, html, LitElement, property, query } from "lit-element";
