+++ web-components/src/components/chip/Chip.stories.ts
@@ -2,8 +2,8 @@ import { withA11y } from "@storybook/addon-a11y";
 import { boolean, number, select, text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-html";
 import { badgeColor, BarType, iconSamples } from "@/utils/enums";
-import "../icon/Icon";
-import "./Chip";
+import "@/components/icon/Icon";
+import "@/components/chip/Chip";
 
 export default {
   title: "Chip",
