+++ web-components/src/components/chip/Chip.test.ts
@@ -1,8 +1,8 @@
-import "../icon/Icon";
+import "@/components/icon/Icon";
 import { Key } from "@/constants";
 import { elementUpdated, fixture, fixtureCleanup, html } from "@open-wc/testing-helpers";
-import "./Chip";
-import { Chip } from "./Chip";
+import "@/components/chip/Chip";
+import { Chip } from "@/components/chip/Chip";
 
 const fixtureFactory = async (): Promise<Chip> => {
   return await fixture(html`
