+++ web-components/src/components/combobox/ComboBox.stories.ts
@@ -1,5 +1,5 @@
-import "./ComboBox";
-import "../icon/Icon";
+import "@/components/combobox/ComboBox";
+import "@/components/icon/Icon";
 import { comboBoxObjectOptions, comboBoxOptions } from "@/[sandbox]/sandbox.mock";
 import { withA11y } from "@storybook/addon-a11y";
 import { boolean, text, withKnobs } from "@storybook/addon-knobs";
