+++ web-components/src/components/editable-textfield/EditableTextfield.test.ts
@@ -1,7 +1,7 @@
 import { elementUpdated, fixture, fixtureCleanup, html } from "@open-wc/testing-helpers";
-import { Input } from "../input/Input";
-import "./EditableTextfield";
-import { EditableTextfield } from "./EditableTextfield";
+import { Input } from "@/components/input/Input";
+import "@/components/editable-textfield/EditableTextfield";
+import { EditableTextfield } from "@/components/editable-textfield/EditableTextfield";
 
 const validInputs = {
   text: "text",
