+++ web-components/src/components/icon/Icon.ts
@@ -12,7 +12,7 @@ import getColorValue from "@momentum-ui/utils/lib/getColorValue";
 import { customElement, html, LitElement, property } from "lit-element";
 import { classMap } from "lit-html/directives/class-map";
 import { styleMap } from "lit-html/directives/style-map";
-import "../button/Button";
+import "@/components/button/Button";
 import styles from "./scss/module.scss";
 
 export const iconSize = ["16", "20", "36", "56", 16, 20, 36, 56] as const;
