+++ web-components/src/components/input/Input.stories.ts
@@ -1,9 +1,17 @@
 import { withA11y } from "@storybook/addon-a11y";
 import { boolean, select, text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-html";
-import "../icon/Icon";
-import "./Input";
-import { containerSize, iconNames, iconPosition, inputShape, inputSize, inputType, nestedLevel } from "./Input";
+import "@/components/icon/Icon";
+import "@/components/input/Input";
+import {
+  containerSize,
+  iconNames,
+  iconPosition,
+  inputShape,
+  inputSize,
+  inputType,
+  nestedLevel
+} from "@/components/input/Input";
 
 export default {
   title: "Input",
