+++ web-components/src/components/label/Label.stories.ts
@@ -1,8 +1,8 @@
 import { withA11y } from "@storybook/addon-a11y";
 import { text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-html";
-import "../input/Input";
-import "./Label";
+import "@/components/input/Input";
+import "@/components/label/Label";
 
 export default {
   title: "Label",
