+++ web-components/src/components/list/List.ts
@@ -6,7 +6,7 @@
  *
  */
 
-import { ListItem } from "../list/ListItem";
+import { ListItem } from "@/components/list/ListItem";
 import { Key } from "@/constants";
 import { RovingTabIndexMixin } from "@/mixins";
 import reset from "@/wc_scss/reset.scss";
