+++ web-components/src/components/menu-overlay/MenuOverlay.stories.ts
@@ -1,12 +1,12 @@
-import "../button/Button";
-import "../checkbox/Checkbox";
-import "../input/Input";
-import "../theme/Theme";
+import "@/components/button/Button";
+import "@/components/checkbox/Checkbox";
+import "@/components/input/Input";
+import "@/components/theme/Theme";
 import { withA11y } from "@storybook/addon-a11y";
 import { boolean, select, text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-html";
-import "./MenuOverlay";
-import { menuOverlayPlacement, menuOverlaySize } from "./MenuOverlay";
+import "@/components/menu-overlay/MenuOverlay";
+import { menuOverlayPlacement, menuOverlaySize } from "@/components/menu-overlay/MenuOverlay";
 
 export default {
   title: "Menu Overlay",
