+++ web-components/src/components/menu-overlay/MenuOverlay.test.ts
@@ -1,5 +1,5 @@
-import "../button/Button";
-import "./MenuOverlay";
+import "@/components/button/Button";
+import "@/components/menu-overlay/MenuOverlay";
 import { Key } from "@/constants";
 import {
   defineCE,
@@ -11,7 +11,7 @@ import {
   oneEvent
 } from "@open-wc/testing-helpers";
 import { html, LitElement, PropertyValues } from "lit-element";
-import { MenuOverlay, MenuOverlayElement, OverlaySizes } from "./MenuOverlay";
+import { MenuOverlay, MenuOverlayElement, OverlaySizes } from "@/components/menu-overlay/MenuOverlay";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnyConstructor<A = LitElement> = new (...args: any[]) => A;
