+++ web-components/src/components/radio/RadioGroup.ts
@@ -6,7 +6,7 @@
  *
  */
 
-import { Radio } from "../radio/Radio";
+import { Radio } from "@/components/radio/Radio";
 import { Key } from "@/constants";
 import { RovingTabIndexMixin } from "@/mixins";
 import reset from "@/wc_scss/reset.scss";
