+++ web-components/src/components/tabs/Tabs.ts
@@ -6,11 +6,11 @@
  *
  */
 
-import "../icon/Icon";
-import "../menu-overlay/MenuOverlay";
+import "@/components/icon/Icon";
+import "@/components/menu-overlay/MenuOverlay";
 import { Key } from "@/constants";
 import { ResizeMixin, RovingTabIndexMixin } from "@/mixins";
-import { uuid } from "@/utils/helpers";
+import { nanoid } from "nanoid";
 import reset from "@/wc_scss/reset.scss";
 import {
   customElement,
@@ -211,7 +211,7 @@ export class Tabs extends ResizeMixin(RovingTabIndexMixin(LitElement)) {
     }
 
     tabs.forEach((tab, index) => {
-      const id = uuid();
+      const id = nanoid();
 
       tab.setAttribute("id", id);
       tab.setAttribute("aria-controls", id);
