+++ web-components/src/components/taskitem/TaskItem.ts
@@ -6,8 +6,8 @@
  *
  */
 
-import "../badge/Badge";
-import "../icon/Icon";
+import "@/components/badge/Badge";
+import "@/components/icon/Icon";
 import reset from "@/wc_scss/reset.scss";
 import { customElement, html, LitElement, property } from "lit-element";
 import { nothing } from "lit-html";
